Switches the Goals Page and the Compete Page as we talked about at the meeting
