+++ pubspec.yaml
@@ -14,7 +14,7 @@ description: The app for Calvin University&#x27;s Grassroots Green group.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.2 <3.0.0"
 
 dependencies:
   flutter:
